Edit unRAID installation guide
Changed to only clone once and and pull in a schedule. This will allow users to only need to grab changes whenever there are new commits. If there are no new commits, then there is no need to remove the files and re-download them.
Also added section for global git configuration, since the git pull may fail without this configuration.